// src/enums/types.ts
export enum EType {
    BOOLEAN = 'boolean',
    DATE = 'date',
    DATE_TIME = 'dateTime',
    TIME = 'time',
    EMAIL = 'email',
    ID = 'id',
    IP = 'ip',
    NULL = 'null',
    UNDEFINED = 'undefined',
    NUMBER = 'number',
    REGEX ='regex',
    WORD ='word',
    SENTENCE = 'sentence',
    URL = 'url',
    OBJECT = 'object',
    ARRAY = 'array',
    CUSTOM = 'custom'
}